+++ src/main/java/com/me/bootSecurityJPAExample/SampleApplication.java
@@ -1,4 +1,4 @@
-package com.me.sample;
+package com.me.bootSecurityJPAExample;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
