+++ src/main/java/com/me/bootSecurityJPAExample/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.me.sample.config;
+package com.me.bootSecurityJPAExample.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 
 @EnableTransactionManagement
-@EnableJpaRepositories(basePackages = "com.me.sample.repository")
+@EnableJpaRepositories(basePackages = "com.me.bootSecurityJPAExample.repository")
 @EnableJpaAuditing
 @Configuration
 public class JpaConfig {
