+++ src/test/java/com/me/bootSecurityJPAExample/SampleApplicationTests.java
@@ -1,4 +1,4 @@
-package com.me.sample;
+package com.me.bootSecurityJPAExample;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
